match the interface of pyro
